Change structure to hide functions
Create Library class and add methods to add and remove books from it.
Create ScreenController IIFE with functions that reflect what's shown on
screen. In it the library array is created that hosts the 'database'.
Move every function that does something with display in it. Do a few
changes to the functions to fix the fact that the book array doesn't
exist in the same scope. Fix redundant way of finding id of the book
that's going to be removed.